Feat: 뷰 닫기 버튼 분리 및 주석 추가
뷰 닫기 버튼이 중복돼서 분리 및 추가 주석 추가
