+++ ttubeokAR/ttubeokAR/PlaceSelect/PlaceSelectViewFlow/PlaceRegistrationView.swift
@@ -21,6 +21,10 @@ struct PlaceRegistrationView: View {
                 .navigationDestination(isPresented: $showSelectPlaceView) {
                     WhichSelectPlaceView()
                 }
+            //TODO: - 버튼 커스텀하여 사용하기
+            /**
+             루트뷰로 나가기 위해 NavigationStack으로 뷰 전환 시 생성되는 버튼 사용하지 말것!
+             */
         }
     }
     
@@ -32,8 +36,7 @@ struct PlaceRegistrationView: View {
             ZStack(alignment: .top) {
                 backgroundView
                 blackOpacityView
-                closeButton
-                    .position(x: geometry.size.width * 0.1)
+                CloseCancelButton(lastedSelectedTab: lastedSelectedTab)
                 TitleView(
                     titleText: "1분만에 장소를 \n 등록해보세요",
                     highlightText: "장소",
@@ -73,28 +76,9 @@ struct PlaceRegistrationView: View {
                 .font(.sandol(type: .medium, size: 20))
                 .multilineTextAlignment(.center)
                 .foregroundStyle(Color.white)
-                .frame(maxWidth: 335, maxHeight: 55)
+                .frame(maxWidth: 335, maxHeight: 42)
                 .background((RoundedRectangle(cornerRadius: 19).fill(Color.primary03)))
         })
-    }
-    
-    private var closeButton: some View {
-        HStack {
-            Button(action: {
-                changeRootViewToMainView(selectedTab: lastedSelectedTab)
-            }){
-                Icon.closeView.image
-                    .resizable()
-                    .frame(maxWidth: 11, maxHeight: 18)
-            }
-        }
-        .frame(maxWidth: 100)
-        .padding(.top, 30)
-    }
-    
-    private func changeRootViewToMainView(selectedTab: Int) {
-        let appDelegate = UIApplication.shared.delegate as? AppDelegate
-        appDelegate?.changeRootViewController(UIHostingController(rootView: MainViewControl(selectedTab: selectedTab).environmentObject(SharedTabInfo())),animated: true)
     }
 }
