+++ ttubeokAR/ttubeokAR/SettingGroup/AppDelegate.swift
@@ -28,8 +28,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         KakaoSDK.initSDK(appKey: "1c831262e8deaf4f7823434057f15384")
         
         // Create the SwiftUI view that provides the window contents.
-         /*MainViewControl().environmentObject(SharedTabInfo())*/
-        let contentView = LoginViewCycle()
+        let contentView = MainViewControl().environmentObject(SharedTabInfo())
         // Use a UIHostingController as window root view controller.
         let window = UIWindow(frame: UIScreen.main.bounds)
         window.rootViewController = UIHostingController(rootView: contentView)
